feat(bot)!: use slash commands and context menus
Migrate everything to slash commands, except for `rotate` which is a mixture of message context
menus and slash commands, and `profile` which is now a user context menu.

BREAKING CHANGE: The old way of using commands is no longer supported, so the `prefix` command has
been removed. The `hktb` and `iwmelc` commands have also been merged into a single `meme` command.